requirements.txt
selenium
bs4
python-telegram-bot
redis